+++ ui-test/tests/Icons.jsx
@@ -20,7 +20,10 @@ const TestIcons = React.createClass({
                 <CardContent>
                     <h5>Basic Icons with Fallback tooltips</h5>
                     <Icon name="edit"/>
-                    <Icon name="delete"/>
+                    <Icon name="remove"/>
+                    <Icon name="save"/>
+                    <Icon name="confirm"/>
+                    <Icon name="cancel"/>
                     <Icon name="arrow_firstpage"/>
                     <Icon name="arrow_prevpage"/>
                     <Icon name="arrow_nextpage"/>
@@ -29,13 +32,17 @@ const TestIcons = React.createClass({
                     <Icon name="expand_more"/>
                     <Icon name="expand_less"/>
                     <Icon name="menu_more"/>
+                    <Icon name="adjust"/>
                     <Icon name="filter"/>
                     <Icon name="sort"/>
                     <Icon name="hide"/>
+                    <Icon name="settings"/>
                     <Icon name="access_forbidden"/>
+                    <Icon name="delete"/>
                     <h5>Other Icons</h5>
                     <Icon name="widgets" tooltip="no tooltip on the next (canonical) icon"/>
                     <Icon name="access_forbidden" tooltip={false}/>
+                    <Icon name="non_existing_icon_name" tooltip="Icon with non existent icon name"/>
                 </CardContent>
             </Card>
         );